requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
asyncpg==0.29.0
PyJWT==2.8.0
bcrypt==4.1.3
pydantic[email]==2.7.0
